[🚚rename]: MyPlant => Plant로 폴더명 수정
