+++ src/pages/Plant/AIRecommend/AIRecommend.tsx
