+++ src/pages/Plant/AIRecommend/SliderContainer/PageIndicator.tsx
