+++ src/pages/Plant/AIRecommend/SliderContainer/PageTitle.tsx
