+++ src/pages/Plant/AIRecommend/SliderContainer/PlantGraph/GraphDefaultOptions.ts
