+++ src/pages/Plant/AIRecommend/SliderContainer/PlantGraph/GraphLegend.tsx
