+++ src/pages/Plant/AIRecommend/SliderContainer/PlantImage/PlantImage.tsx
