+++ src/pages/Plant/AIRecommend/SliderContainer/PlantName/PlantName.tsx
