+++ src/pages/Plant/AIRecommend/SliderContainer/SliderContainer.tsx
