+++ src/pages/Plant/AIRecommend/components/InputField.tsx
@@ -1,5 +1,5 @@
-import { Label } from '@/components/ui/label';
 import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
 
 interface InputFieldProps {
   label: string;
