+++ src/pages/Plant/AIRecommend/constants/titles.ts
