+++ src/pages/Plant/AIRecommend/type.ts
