+++ src/pages/Plant/PlantDetail/PlantDetail.tsx
