+++ src/pages/PurchaseHistoryPage/type.ts
@@ -5,12 +5,6 @@ interface PurchaseItems {
   paymentTime: string;
 }
 
-interface MenuItem {
-  name: string;
-  url: string;
-}
-
 export interface PurchaseHistoryPageProps {
-  menuItems: MenuItem[];
   purchaseItems: PurchaseItems[];
 }
